fix: Next.js 16 Turbopack 빌드 오류 수정
- 13개 actions.ts barrel에서 'use server' 제거 (Turbopack은 'use server' 파일에서 async function 직접 export만 허용)
- options-chain/index.ts에서 server-only export 분리 (yahoo-finance2 Node.js 의존성이 client 번들에 포함되는 문제)
- session/index.ts에서 getCurrentUser barrel 제외 (next/headers client 번들 포함 문제)
- cleanupExpiredSessionsAction에서 class를 별도 파일로 분리 ('use server' 파일 non-function export 금지)
- 테스트 mock 경로를 새 import 경로에 맞게 업데이트
- entities/CLAUDE.md, features/CLAUDE.md에 규칙 문서화
- eslint.config.mjs: RSC pages가 server-only lib 경로를 import할 수 있도록 eslint 예외 추가
