+++ src/app/account/delete/page.tsx
@@ -4,7 +4,7 @@ import { Suspense } from 'react';
 import { redirect } from 'next/navigation';
 import { AuthCardShell } from '@/shared/ui/auth/AuthCardShell';
 import { DeleteAccountConfirm } from '@/features/account-delete';
-import { getCurrentUser } from '@/entities/session';
+import { getCurrentUser } from '@/entities/session/lib/getCurrentUser';
 import { SITE_NAME, SITE_URL } from '@/shared/lib/seo';
 
 // noindex 페이지에도 canonical/og:url을 명시한다 (login/signup 정책과 일관).
