+++ src/entities/api-key/actions.ts
@@ -1,8 +1,3 @@
-// barrel을 server boundary로 강제하여 client 실수 import를 컴파일 타임에 차단.
-// 개별 action 파일도 각각 'use server'를 선언하지만, barrel에도 명시해야
-// Next.js가 이 진입점 자체를 server-only로 인식한다.
-'use server';
-
 export { saveApiKeyAction } from './actions/saveApiKeyAction';
 export { deleteApiKeyAction } from './actions/deleteApiKeyAction';
 export { getRegisteredProvidersAction } from './actions/getRegisteredProvidersAction';
