+++ src/entities/bars/actions.ts
@@ -1,3 +1,1 @@
-'use server';
-
 export { getBarsAction } from './actions/getBarsAction';
