+++ src/entities/chat-message/actions.ts
@@ -1,4 +1,2 @@
-'use server';
-
 export { chatAction } from './actions/chatAction';
 export { getRemainingTokensAction } from './actions/getRemainingTokensAction';
