+++ src/entities/inquiry/actions.ts
@@ -1,3 +1,1 @@
-'use server';
-
 export { submitContactAction } from './actions/submitContactAction';
