+++ src/entities/market-summary/actions.ts
@@ -1,3 +1,1 @@
-'use server';
-
 export { getMarketSummaryAction } from './actions/getMarketSummaryAction';
