+++ src/entities/options-chain/actions.ts
@@ -1,5 +1,3 @@
-'use server';
-
 export {
     submitOptionsAnalysisAction,
     pollOptionsAnalysisAction,
