+++ src/entities/sector-signal/actions.ts
@@ -1,3 +1,1 @@
-'use server';
-
 export { getSectorSignalsAction } from './actions/getSectorSignalsAction';
