+++ src/entities/session/actions.ts
@@ -1,4 +1,2 @@
-'use server';
-
 export { currentUserAction } from './actions/currentUserAction';
 export { cleanupExpiredSessionsAction } from './actions/cleanupExpiredSessionsAction';
