+++ src/entities/ticker/actions.ts
@@ -1,4 +1,2 @@
-'use server';
-
 export { getAssetInfoAction } from './actions/getAssetInfoAction';
 export { searchTickerAction } from './actions/searchTickerAction';
