+++ src/entities/user-tier/actions.ts
@@ -1,3 +1,1 @@
-'use server';
-
 export { getUserTierAction } from './actions/getUserTierAction';
