+++ src/features/CLAUDE.md
@@ -14,6 +14,14 @@ features는 `entities/`와 `shared/`만 import 가능. **상위 레이어(widget
 
 이 예외는 ESLint `from: 'features', allow: ['features', ...]`로 관리됨. Phase 7 cleanup 시 해소 가능.
 
+## `'use server'` 규칙
+
+`actions.ts` barrel 파일에 `'use server'`를 선언하면 **안 된다.**
+Next.js 16 Turbopack은 `'use server'` 파일에서 async function 직접 export만 허용하며,
+re-export 문은 빌드 오류를 유발한다. 개별 action 파일에서만 `'use server'`를 선언한다.
+
+자세한 규칙은 `src/entities/CLAUDE.md` § `'use server'` 규칙 참조.
+
 ## 슬라이스 구조
 
 ```
@@ -21,7 +29,7 @@ features/<name>/
 ├── ui/               UI 컴포넌트
 ├── model/            (선택) Context + Provider
 ├── hooks/            React hooks
-├── actions/          (선택) Server Action wrapper
+├── actions/          (선택) Server Action wrapper (❗ barrel에 'use server' 선언 금지)
 ├── lib/              (선택) 순수 함수
 ├── __tests__/        colocated tests
 └── index.ts          public API barrel
