+++ src/features/auth-oauth/actions.ts
@@ -1,3 +1,1 @@
-'use server';
-
 export { cancelOAuthSignupAction } from './actions/cancelOAuthSignupAction';
